Camera
I wanted to write a function called 'shift', but I realized the
direction in which the audio would shift was ambiguous. I could clarify
which direction does what in a comment, but it will still be
error-prone (because if a caller didn't read the documentation, the call
won't look wrong) and annoying (if the caller always has to negate their
argument).

Using this type class, the implementer can use the direction which makes
the most sense to them, the caller can use the direction which makes the
most sense for them, and it should be obvious from the call site what
the direction means.